DB Schema Set

+++ src/schema.rs
@@ -2,8 +2,8 @@ use serde::Deserialize;
 
 #[derive(Debug, Deserialize, Default)]
 pub struct Pagination {
-    pub offset: Option<i32>,
-    pub limit: Option<i32>,
+    pub offset: Option<i64>,
+    pub limit: Option<i64>,
 }
 
 #[derive(Debug, Deserialize)]
